Increase speed of stylesheet matching.
On the dry run when matching widget against different selectors in stylesheet we remember the selectors that match atleast partially (tagname, id & structure) and remember them in array. Next time we run stylesheet lookup routine it doesnt look through all of them. It only matches against previously remembered "possible" stylesheet rules.

The speed can be compared to what it used to be in old times with caching - the fastest version i had.